# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for FECh Plenario
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sessions/{id}/speaking-turns:
    get:
      operationId: listSpeakingTurns
      tags: [speaking]
      summary: Listar la cola de palabras de una sesión
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cola de palabras
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SpeakingTurn"
    post:
      operationId: requestSpeakingTurn
      tags: [speaking]
      summary: Pedir/agregar una palabra individual a la cola
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpeakingTurnInput"
      responses:
        "201":
          description: Palabra agregada a la cola
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingTurn"
        "400":
          description: Datos inválidos
        "409":
          description: Ya tienes una palabra en cola para este punto
  /sessions/{id}/speaking-turns/collective:
    post:
      operationId: createCollectiveSpeakingTurn
      tags: [speaking]
      summary: Armar una palabra colectiva por Unidad Académica (consejeros)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpeakingTurnCollectiveInput"
      responses:
        "201":
          description: Palabra colectiva creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingTurn"
        "400":
          description: No hay consejeros presentes de esa unidad
  /sessions/{id}/speaking-round:
    post:
      operationId: toggleSpeakingRound
      tags: [speaking]
      summary: Abrir o cerrar la ronda de palabras de una sesión (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpeakingRoundInput"
      responses:
        "200":
          description: Estado de la ronda de palabras actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          description: Datos inválidos
        "404":
          description: Sesión no encontrada
  /sessions/{id}/speaking-turns/reorder:
    post:
      operationId: reorderSpeakingTurns
      tags: [speaking]
      summary: Reordenar la cola de palabras
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpeakingReorderInput"
      responses:
        "200":
          description: Cola reordenada
  /speaking-turns/{turnId}:
    patch:
      operationId: updateSpeakingTurn
      tags: [speaking]
      summary: Ajustar duración o punto de una palabra (admin)
      parameters:
        - name: turnId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpeakingTurnUpdate"
      responses:
        "200":
          description: Palabra actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingTurn"
        "404":
          description: Palabra no encontrada
    delete:
      operationId: deleteSpeakingTurn
      tags: [speaking]
      summary: Quitar una palabra de la cola (admin)
      parameters:
        - name: turnId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Palabra eliminada
  /speaking-turns/{turnId}/control:
    post:
      operationId: controlSpeakingTurn
      tags: [speaking]
      summary: Controlar el turno (otorgar/iniciar/pausar/finalizar)
      parameters:
        - name: turnId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpeakingTurnControlInput"
      responses:
        "200":
          description: Turno actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingTurn"
        "404":
          description: Palabra no encontrada
...
components:
  schemas:
...
    Session:
      type: object
      required: [id, title, status, sessionCode, speakingRoundOpen, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        location:
          type: ["string", "null"]
        scheduledAt:
          type: ["string", "null"]
        status:
          type: string
          enum: [abierta, cerrada]
        sessionCode:
          type: string
        meetingLink:
          type: ["string", "null"]
        actaObjectPath:
          type: ["string", "null"]
        actaFileName:
          type: ["string", "null"]
        speakingRoundOpen:
          type: boolean
        speakingRoundAgendaPointId:
          type: ["integer", "null"]
        officialStartAt:
          type: ["string", "null"]
          description: Momento de la apertura oficial. Nulo si nunca se abrió oficialmente.
        officialEndAt:
          type: ["string", "null"]
          description: Momento del cierre. Nulo si sigue en curso o si nunca se abrió oficialmente.
        createdAt:
          type: string
...
    SpeakingTurnParticipant:
      type: object
      required: [userId, displayName]
      properties:
        userId:
          type: integer
        displayName:
          type: string
        faculty:
          type: ["string", "null"]
    SpeakingTurn:
      type: object
      required: [id, sessionId, category, kind, status, position, durationSeconds, elapsedSeconds, label, participants, createdAt]
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        agendaPointId:
          type: ["integer", "null"]
        category:
          type: string
          enum: [pleno, base]
        kind:
          type: string
          enum: [individual, colectiva]
        status:
          type: string
          enum: [en_cola, hablando, finalizada]
        position:
          type: integer
        durationSeconds:
          type: integer
        elapsedSeconds:
          type: integer
        startedAt:
          type: ["string", "null"]
        faculty:
          type: ["string", "null"]
        userId:
          type: ["integer", "null"]
        label:
          type: string
        participants:
          type: array
          items:
            $ref: "#/components/schemas/SpeakingTurnParticipant"
        createdAt:
          type: string
    SpeakingTurnInput:
      type: object
      required: [category]
      properties:
        category:
          type: string
          enum: [pleno, base]
        agendaPointId:
          type: ["integer", "null"]
        userId:
          type: ["integer", "null"]
        speakerName:
          type: ["string", "null"]
        durationSeconds:
          type: ["integer", "null"]
    SpeakingTurnCollectiveInput:
      type: object
      required: [faculty]
      properties:
        faculty:
          type: string
        agendaPointId:
          type: ["integer", "null"]
        userIds:
          type: array
          items:
            type: integer
        secondsPerSpeaker:
          type: ["integer", "null"]
        durationSeconds:
          type: ["integer", "null"]
    SpeakingTurnUpdate:
      type: object
      properties:
        durationSeconds:
          type: integer
        agendaPointId:
          type: ["integer", "null"]
    SpeakingTurnControlInput:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [grant, start, pause, finish]
    SpeakingReorderInput:
      type: object
      required: [orderedIds]
      properties:
        orderedIds:
          type: array
          items:
            type: integer
    SpeakingRoundInput:
      type: object
      required: [open]
      properties:
        open:
          type: boolean
        agendaPointId:
          type: ["integer", "null"]
          description: Punto de tabla al que se limita la ronda (obligatorio al abrir)
...